+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - PR #357 Improved error checking on memory allocates
 - PR #358 Added improved hash algorithm to reduce collisions
 - PR #369 Reconfigure C++ source directory structure
+- PR #370 Updated memsize() to include pointer memory too
 
 ## Bug Fixes